added maven and bouncy castle created basic structure for program Does a very basic proof of work

+++ src/main/java/com/ssd/Block.java
@@ -7,7 +7,7 @@ public class Block{
     public String previousHash;
     private String data;
     private long timeStamp;
-    private int nonce;
+    private long nonce;
 
     public Block( String data, String previousHash) {
         this.previousHash = previousHash;
